Do not error on stderr output, but do error on error exit (fix #18)
Remove the call to _onScriptError when there is stderr output. Stderr output does not in general indicate an error.

Use the inherit option for stdio rather than pipe in the call to spawn, so that the child's output is visible. This is what one will expect from general shell scripts.

In _onScriptComplete, call _onScriptError when the exit status is non-zero.

Avoid a trailing space on the end of the name of the script shown in messages.